Re-org gossip validation fns
